+++ include/Basic/Color3.cuh
@@ -18,16 +18,16 @@ namespace renderer {
      *   生成每个分量都在指定范围内的随机颜色
      */
     typedef struct Color3 {
-        double r, g, b;
+        float r, g, b;
 
         // ====== 成员访问 ======
-        __host__ __device__ double operator[](size_t index) const {
+        __host__ __device__ float operator[](size_t index) const {
             switch (index) {
                 case 0: return r; case 1: return g;
                 case 2: return b; default: return r;
             }
         }
-        __host__ __device__ double & operator[](size_t index) {
+        __host__ __device__ float & operator[](size_t index) {
             switch (index) {
                 case 0: return r; case 1: return g;
                 case 2: return b; default: return r;
@@ -72,39 +72,39 @@ namespace renderer {
         }
 
         //数乘除
-        __host__ __device__ void operator*=(double num) {
+        __host__ __device__ void operator*=(float num) {
             for (size_t i = 0; i < 3; i++) {
                 operator[](i) *= num;
             }
         }
-        __host__ __device__ Color3 operator*(double num) const {
+        __host__ __device__ Color3 operator*(float num) const {
             Color3 ret = *this; ret *= num; return ret;
         }
-        __host__ __device__ friend Color3 operator*(double num, const Color3 & obj) {
+        __host__ __device__ friend Color3 operator*(float num, const Color3 & obj) {
             return obj * num;
         }
-        __host__ __device__ void operator/=(double num) {
+        __host__ __device__ void operator/=(float num) {
             for (size_t i = 0; i < 3; i++) {
                 operator[](i) /= num;
             }
         }
-        __host__ __device__ Color3 operator/(double num) const {
+        __host__ __device__ Color3 operator/(float num) const {
             Color3 ret = *this; ret /= num; return ret;
         }
-        __host__ __device__ friend Color3 operator/(double num, const Color3 & obj) {
+        __host__ __device__ friend Color3 operator/(float num, const Color3 & obj) {
             return obj / num;
         }
 
         //颜色转换
-        __device__ uchar4 castToUchar4(double gamma = 2.0) const {
+        __device__ uchar4 castToUchar4(float gamma = 2.0) const {
             //伽马校正
-            const double power = 1.0 / gamma;
-            const double _r = pow(r, power);
-            const double _g = pow(g, power);
-            const double _b = pow(b, power);
+            const float power = 1.0f / gamma;
+            const float _r = pow(r, power);
+            const float _g = pow(g, power);
+            const float _b = pow(b, power);
 
             //将[0.0, 1.0]的颜色值映射到[0, 255]
-            static constexpr Range colorRange{0.0, 0.999};
+            static constexpr Range colorRange{0.0f, 0.999f};
             const auto intR = static_cast<Uint8>(256 * colorRange.clamp(_r));
             const auto intG = static_cast<Uint8>(256 * colorRange.clamp(_g));
             const auto intB = static_cast<Uint8>(256 * colorRange.clamp(_b));
@@ -114,14 +114,14 @@ namespace renderer {
         }
 
         // ====== 随机颜色生成 ======
-        static Color3 randomColor(double componentMin = 0.0, double componentMax = 1.0) {
+        static Color3 randomColor(float componentMin = 0.0f, float componentMax = 1.0f) {
             Color3 ret{};
             for (size_t i = 0; i < 3; i++) {
                 ret[i] = RandomGenerator::randomDouble(componentMin, componentMax);
             }
             return ret;
         }
-        __device__ static Color3 randomColor(curandState * state, double componentMin = 0.0, double componentMax = 1.0) {
+        __device__ static Color3 randomColor(curandState * state, float componentMin = 0.0f, float componentMax = 1.0f) {
             Color3 ret{};
             for (size_t i = 0; i < 3; i++) {
                 ret[i] = RandomGenerator::randomDouble(state, componentMin, componentMax);
