+++ include/Material/Metal.cuh
@@ -10,7 +10,7 @@ namespace renderer {
      */
     typedef struct Metal {
         Color3 albedo;
-        double fuzz;
+        float fuzz;
 
         __device__ bool scatter(curandState * state, const Ray & in, const HitRecord & record, Color3 & attenuation, Ray & out) const {
             //计算反射光线方向向量（单位向量）
@@ -19,8 +19,8 @@ namespace renderer {
             Vec3 reflectDirection = (v - 2 * Vec3::dot(v, n) * n).unitVector();
 
             //应用反射扰动：在距离物体表面1单位处随机选取单位向量和反射向量相加，形成随机扰动
-            if (fuzz > 0.0) {
-                reflectDirection += fuzz * Vec3::randomSpaceVector(state, 1.0);
+            if (fuzz > 0.0f) {
+                reflectDirection += fuzz * Vec3::randomSpaceVector(state, 1.0f);
             }
 
             //构建反射光线，光线的时间属性不随传播而改变
@@ -28,7 +28,7 @@ namespace renderer {
             attenuation = albedo;
 
             //防止由于计算精度导致out方向向内。使用点积检查反射光线是否和物体外法线的同侧，仅当二者同侧时有效
-            return Vec3::dot(out.direction, record.normalVector) > 0.0;
+            return Vec3::dot(out.direction, record.normalVector) > 0.0f;
         }
     } Metal;
 }
